separate file has been made for schema and resolver
