+++ app.js
@@ -1,148 +1,20 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const {graphqlHTTP} = require('express-graphql');
-const { buildSchema } = require('graphql');
 const mongoose = require('mongoose');
-const bcrypt = require('bcryptjs');
-
-const Event = require('./model/event');
-const User = require('./model/user');
+const graphqlSchema = require("./graphql/schema/index"); 
+const graphqlResolver = require("./graphql/resolvers/index"); 
 
 const app = express();
 
 app.use(bodyParser.json());
 
-const user = userId => {
-    return User.findById(userId)
-    .then(user => {
-        return { ...user._doc, _id: user.id, createdEvents:events.bind(this, user._doc.createdEvents) }
-    })
-    .catch(err => {
-        throw err; 
-    })
-}
-
-const events = eventIds => {
-    return Event.find({ _id: {$in: eventIds}})
-    .then(events => {
-        return events.map(event => {
-            return { ...event._doc, _id: event.id, creator: user.bind(this, event._doc.creator)}
-        })
-    })
-    .catch(err => {
-        throw err; 
-    })
-}
 
 app.use(
   '/graphql',
   graphqlHTTP({
-    schema: buildSchema(`
-        type Event {
-          _id: ID!
-          title: String!
-          description: String!
-          price: Float!
-          date: String!
-          creator: User!
-        }
-        type User {
-          _id: ID!
-          email: String!
-          password: String
-          createdEvents: [Event!]
-        }
-        input EventInput {
-          title: String!
-          description: String!
-          price: Float!
-          date: String!
-        }
-        input UserInput {
-          email: String!
-          password: String!
-        }
-        type RootQuery {
-            events: [Event!]!
-        }
-        type RootMutation {
-            createEvent(eventInput: EventInput): Event
-            createUser(userInput: UserInput): User
-        }
-        schema {
-            query: RootQuery
-            mutation: RootMutation
-        }
-    `),
-    rootValue: {
-      events: () => {
-        return Event.find()
-          .then(events => {
-            return events.map(event => {
-              return {
-                ...event._doc,
-                 _id: event.id , 
-                 creator: user.bind(this, event._doc.creator)
-                };
-            });
-          })
-          .catch(err => {
-            throw err;
-          });
-      },
-      createEvent: args => {
-        const event = new Event({
-          title: args.eventInput.title,
-          description: args.eventInput.description,
-          price: +args.eventInput.price,
-          date: new Date(args.eventInput.date),
-          creator: '603488d1d2b05fff018adfe1'
-        });
-        let createdEvent;
-        return event
-          .save()
-          .then(result => {
-            createdEvent = { ...result._doc, _id: result._doc._id.toString() };
-            return User.findById('603488d1d2b05fff018adfe1');
-          })
-          .then(user => {
-            if (!user) {
-              throw new Error('User not found.');
-            }
-            user.createdEvents.push(event);
-            return user.save();
-          })
-          .then(result => {
-            return createdEvent;
-          })
-          .catch(err => {
-            console.log(err);
-            throw err;
-          });
-      },
-      createUser: args => {
-        return User.findOne({ email: args.userInput.email })
-          .then(user => {
-            if (user) {
-              throw new Error('User exists already.');
-            }
-            return bcrypt.hash(args.userInput.password, 12);
-          })
-          .then(hashedPassword => {
-            const user = new User({
-              email: args.userInput.email,
-              password: hashedPassword
-            });
-            return user.save();
-          })
-          .then(result => {
-            return { ...result._doc, password: null, _id: result.id };
-          })
-          .catch(err => {
-            throw err;
-          });
-      }
-    },
+    schema: graphqlSchema,
+    rootValue: graphqlResolver,
     graphiql: true
   })
 );
